Reorganize constants and module layout
