+++ curve_dao/vote_utils.py
@@ -1,6 +1,4 @@
 import json
-import os
-import sys
 import warnings
 from typing import Dict, List, Tuple
 
@@ -8,19 +6,11 @@ import ape
 import requests
 from ape.logging import logger
 
-from curve_dao.decoder_utils import decode_input
+from .addresses import get_dao_voting_contract
+from .decoder_utils import decode_input
+from .ipfs import get_ipfs_hash_from_description
 
 warnings.filterwarnings("ignore")
-
-
-warnings.filterwarnings("ignore")
-
-CONVEX_VOTERPROXY = "0x989AEB4D175E16225E39E87D0D97A3360524AD80"
-DAO_VOTING_CONTRACTS = {
-    "ownership": "0xE478de485ad2fe566d49342Cbd03E49ed7DB3356",
-    "parameter": "0xbcff8b0b9419b9a88c44546519b1e909cf330399",
-    "emergency": "0x1115c9b3168563354137cdc60efb66552dd50678",
-}
 
 
 def prepare_vote_script(target: Dict, actions: List[Tuple]) -> str:
@@ -56,44 +46,6 @@ def prepare_vote_script(target: Dict, actions: List[Tuple]) -> str:
     return evm_script
 
 
-def get_ipfs_hash_from_description(description: str):
-    """Uploads vote description to IPFS and returns the IPFS hash.
-
-    NOTE: needs environment variables for infura IPFS access. Please
-    set up an IPFS project to generate project id and project secret!
-    """
-    text = json.dumps({"text": description})
-    response = requests.post(
-        "https://ipfs.infura.io:5001/api/v0/add",
-        files={"file": text},
-        auth=(os.getenv("IPFS_PROJECT_ID"), os.getenv("IPFS_PROJECT_SECRET")),
-    )
-    assert (
-        200 <= response.status_code < 400
-    ), f"POST to IPFS failed: {response.status_code}"
-    return response.json()["Hash"]
-
-
-def get_description_from_ipfs_hash(ipfs_hash: str):
-    response = requests.post(
-        f"https://ipfs.infura.io:5001/api/v0/get?arg={ipfs_hash}",
-        auth=(os.getenv("IPFS_PROJECT_ID"), os.getenv("IPFS_PROJECT_SECRET")),
-    )
-    response.raise_for_status()
-    response_string = response.content.decode("utf-8")
-    json_string = []
-    in_json = False
-    for c in response_string:
-        if c == "{":
-            in_json = True
-        if in_json:
-            json_string.append(c)
-        if c == "}":
-            break
-    json_string = "".join(json_string)
-    return json.loads(json_string)
-
-
 def make_vote(target: Dict, actions: List[Tuple], description: str, vote_creator: str):
     """Prepares EVM script and creates an on-chain AragonDAO vote.
 
@@ -112,9 +64,10 @@ def make_vote(target: Dict, actions: List[Tuple], description: str, vote_creator
     evm_script = prepare_vote_script(target, actions)
     logger.info(f"EVM script: {evm_script}")
 
+    ipfs_hash = get_ipfs_hash_from_description(description)
     tx = aragon.newVote(
         evm_script,
-        f"ipfs:{get_ipfs_hash_from_description(description)}",
+        f"ipfs:{ipfs_hash}",
         False,
         False,
         sender=vote_creator,
@@ -123,8 +76,8 @@ def make_vote(target: Dict, actions: List[Tuple], description: str, vote_creator
     return tx
 
 
-def get_vote_script(vote_id: str, target: str) -> str:
-    voting_contract_address = DAO_VOTING_CONTRACTS[target]
+def get_vote_script(vote_id: str, vote_type: str) -> str:
+    voting_contract_address = get_dao_voting_contract(vote_type)
     voting_contract = ape.project.Voting.at(voting_contract_address)
     vote = voting_contract.getVote(vote_id)
     script = vote["script"]
@@ -210,25 +163,3 @@ def format_fn_inputs(inputs_with_names):
     name, arg = inputs_with_names[-1]
     formatted_args += f"    └─ [bold]{name}[/]: [yellow]{arg}[/]"
     return formatted_args
-
-
-def get_ipfs_hash_from_vote_id(target, vote_id):
-    voting_contract_address = DAO_VOTING_CONTRACTS[target]
-    voting_contract = ape.project.Voting.at(voting_contract_address)
-    snapshot_block = voting_contract.getVote(vote_id)["snapshotBlock"]
-    vote_events = voting_contract.StartVote.query(
-        "voteId",
-        "metadata",
-        start_block=snapshot_block - 1,
-        stop_block=snapshot_block + 1,
-    )
-    vote_row = vote_events.loc[vote_events["voteId"] == vote_id]
-    ipfs_hash = vote_row["metadata"].iloc[0]
-    ipfs_hash = ipfs_hash[5:]
-    return ipfs_hash
-
-
-def get_description_from_vote_id(vote_id, target):
-    ipfs_hash = get_ipfs_hash_from_vote_id(target, vote_id)
-    description = get_description_from_ipfs_hash(ipfs_hash)
-    return description
