+++ tests/test_kill_gauge.py
@@ -1,7 +1,6 @@
 import ape
 import pytest
-
-from curve_dao import CRYPTOSWAP_OWNER_PROXY, CURVE_DAO_OWNERSHIP
+from curve_dao.addresses import CRYPTOSWAP_OWNER_PROXY, CURVE_DAO_OWNERSHIP
 from curve_dao.simulate import simulate
 from curve_dao.vote_utils import make_vote
