+++ tests/test_whitelist.py
@@ -1,7 +1,6 @@
 import ape
 import pytest
-
-from curve_dao import CURVE_DAO_OWNERSHIP
+from curve_dao.addresses import CURVE_DAO_OWNERSHIP
 from curve_dao.modules.smartwallet_checker import (
     SMARTWALLET_CHECKER,
     whitelist_vecrv_lock,
